requirements.txt
DateTime==5.5
filelock==3.14.0
fsspec==2024.5.0
GPUtil==1.4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
pillow==10.3.0
psutil==5.9.8
pytz==2024.1
PyYAML==6.0.1
sympy==1.12
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
triton==2.3.0
typing==3.7.4.3
typing_extensions==4.12.0
zope.interface==6.4.post2
